Ensure ChannelPromise that are used by ChannelOutboundHandler will always complete in the assumed thread.
Motivation:

At the moment we often make the assumption that the ChannelFutureListener attached by an ChannelOutboundHandler will execute in the same thread as the ChannelOutboundHandler by itself. This is not really safe to assume as the user can pass in a ChannelPromise that is tight to another EventExecutor or the previous ChannelOutboundHandler might use a different EventExecutor as the current one and might have created the ChannelPromise.

Modifications:

- If the ChannelPromise is not using the "correct" EventExecutor just create a new one, cascade its result to the original ChannelPromise and pass the new ChannelPromise to the ChannelOutboundHandler. This ensures we execute stuff in the assumed thread.
- Add unit test

Result:

More safe behaviour when executing ChannelOutboundHandler