ChannelPipeline: Add tests for using ChannelHandlerContext and trigger operations
Motivation:

Recently I did spent a lot of time debugging a regression introduced while refactoring code which happend to be an issue with selecting the right handler / context when triggering operations from outside the eventloop. Unfortunally we had no tests for this and so the debugging took quite some time. Let's change that

Modifications:

- Add unit tests to assert correct selection of context / handler

Result:

More testing