Clean up Future/Promises API (#16039)
Motivation:
The generics for the existing futures, promises, and listeners are too
complicated. This complication comes from the existence of
`ChannelPromise` and `ChannelFuture`, which forces listeners to care
about the particular _type_ of future being listened on.

Modification:
* Make the `FutureListener`, the default listener API, retiring the
`GenericFutureListener` since we no longer need to abstract over the
type of the future.
* Change all uses of `ChannelPromise` to `Promise<Void>`.
* Change all uses of `ChannelFuture` to `Future<Void>`.
* Change all uses of `GenericFutureListener` to `FutureListener` as
needed.
* Remove `ChannelFutureListener` and `GenericFutureListener`.
* Remove `ChannelPromise` and `ChannelFuture` — all usages now rely on
the plain `Future` and `Promise` APIs.
* Remove `CompleteChannelFuture`, `SucceededChannelFuture`, and
`FailedChannelFuture`.

Result:
Cleaner generics and more straight forward code.