API: Change method signature to return void when promise is given (#16051)
Motivation:

We usually provided two variants for methods. One that just returns a
Future<...> and one that takes a Promise<...> and returns a Future<...>.
While the first is ok the other one is kind of odd and confusing in the
sense that its not really clear if a listener is better added to the
promse or the future. Let's make the API a bit cleaner and just change
the return type to void if a Promise<...> is provided.

Modifications:

- Change all method signatures from `Future method(..., Promise)` to
`void method(..., Promise)` and so make things more clear and
consistent.

Result:

API cleanup