Fix a problem in the buddy allocator
When releasing parent nodes we were only considering the siblings one level down, not whether any other child of a given node had been claimed.
This could lead to grand-parents getting marked as free, if we returned from a child that had a free sibling.
Now we return whether the given subtree is free and only consider the sibling if so.